Check client-only-vars in determining if signal is supported
